Moved components to subfolder
